feature: move files to match django project structure
