+++ packages/ap-storage/alliance_platform/storage/apps.py
@@ -2,7 +2,7 @@ from alliance_platform.storage.settings import ap_storage_settings
 from django.apps.config import AppConfig
 
 
-class AlliancePlatformCodegenConfig(AppConfig):
+class AlliancePlatformStorageConfig(AppConfig):
     name = "alliance_platform.storage"
     verbose_name = "Alliance Platform Storage"
     label = "alliance_platform_storage"
