+++ packages/ap-storage/alliance_platform/storage/async_uploads/models/models.py
@@ -2,8 +2,6 @@ from __future__ import annotations
 
 from dataclasses import dataclass
 from hashlib import shake_256
-import json
-from json import JSONDecodeError
 import logging
 import mimetypes
 from typing import TYPE_CHECKING
@@ -13,33 +11,27 @@ from typing import cast
 from urllib.parse import urlencode
 
 from alliance_platform.core.auth import resolve_perm_name
-from alliance_platform.storage.base import AsyncUploadStorage
-from alliance_platform.storage.models import AsyncTempFile
-from alliance_platform.storage.registry import AsyncFieldRegistry
-from alliance_platform.storage.registry import default_async_field_registry
+from alliance_platform.storage.async_uploads.registry import default_async_field_registry
+from alliance_platform.storage.async_uploads.storage.base import AsyncUploadStorage
 from allianceutils.middleware import CurrentRequestMiddleware
-from django import forms
 from django.contrib.contenttypes.models import ContentType
-from django.core import validators
 from django.core.exceptions import ValidationError
 from django.db import models
 from django.db import transaction
-from django.db.models import FileField
-from django.db.models import ImageField
 from django.db.models import Model
 from django.db.models import Q
 from django.db.models import signals
 from django.db.models.fields.files import FieldFile
-from django.db.models.fields.files import FileDescriptor
-from django.db.models.fields.files import ImageFieldFile
-from django.forms.widgets import Input
 from django.urls import reverse
-from django.utils.deconstruct import deconstructible
 
 logger = logging.getLogger("alliance_platform.storage")
 
 
 if TYPE_CHECKING:
+    # this is to avoid accidentally importing the incorrect classes that are
+    # only in place for type checking
+    from alliance_platform.storage.async_uploads.models.fields import AsyncFileField as _AsyncFileField
+    from alliance_platform.storage.async_uploads.models.fields import AsyncImageField as _AsyncImageField
 
     class AsyncFileMixinProtocol(Protocol):
         model: type[models.Model]
@@ -51,7 +43,7 @@ if TYPE_CHECKING:
     class FieldFileMixinProtocol(Protocol):
         instance: models.Model
         name: str | None
-        field: AsyncFileField | AsyncImageField
+        field: _AsyncFileField | _AsyncImageField
 
 else:
 
@@ -59,6 +51,12 @@ else:
         pass
 
     class FieldFileMixinProtocol:
+        pass
+
+    class _AsyncFileField:
+        pass
+
+    class _AsyncImageField:
         pass
 
 
@@ -75,7 +73,7 @@ class AsyncFileModelRegistry:
     files.
     """
 
-    fields: list[AsyncFileField | AsyncImageField]
+    fields: list[_AsyncFileField | _AsyncImageField]
     _save_in_progress: list[Model]
 
     def __init__(self):
@@ -83,7 +81,7 @@ class AsyncFileModelRegistry:
         self._save_in_progress = []
 
     @classmethod
-    def register_field(cls, model_class: type[Model], field: AsyncFileField | AsyncImageField):
+    def register_field(cls, model_class: type[Model], field: _AsyncFileField | _AsyncImageField):
         key = "__async_file_model_registry"
         if not hasattr(model_class, key):
             registry = AsyncFileModelRegistry()
@@ -103,7 +101,7 @@ class AsyncFileModelRegistry:
         If the destination filename is too long for the database after ``upload_to`` is applied then the
         file will be truncated by the ``Storage.get_available_name`` method.
 
-        Also see :meth:`~alliance_platform.storage.base.AsyncUploadStorage.move_file`
+        Also see :meth:`~alliance_platform.storage.async_uploads.storage.AsyncUploadStorage.move_file`
 
         This uses as few queries as possible to do the moves and update new values. Only async fields
         that have a temporary key are touched - other fields that have already been moved are left
@@ -197,24 +195,24 @@ _UNSET = object()
 
 
 class AsyncFileMixin(AsyncFileMixinProtocol):
-    """Mixin for file fields that works with :class:`~alliance_platform.storage.base.AsyncUploadStorage` to
+    """Mixin for file fields that works with :class:`~alliance_platform.storage.async_uploads.storage.AsyncUploadStorage` to
     handle uploading directly to external service like S3 or Azure.
 
-    This field works in conjunction with :class:`~alliance_platform.storage.views.GenerateUploadUrlView`. The view will
+    This field works in conjunction with :class:`~alliance_platform.storage.async_uploads.views.GenerateUploadUrlView`. The view will
     generate a URL (eg. a signed URL when using S3) that the frontend can then upload to. Each view is tied
-    to a specific registry which you can specify in ``async_field_registry`` (defaults to :data:`~alliance_platform.storage.registry.default_async_field_registry`).
+    to a specific registry which you can specify in ``async_field_registry`` (defaults to :data:`~alliance_platform.storage.async_uploads.registry.default_async_field_registry`).
 
-    The permissions used by :class:`~alliance_platform.storage.views.GenerateUploadUrlView` can be specified in
+    The permissions used by :class:`~alliance_platform.storage.async_uploads.views.GenerateUploadUrlView` can be specified in
     ``perm_create`` and ``perm_update``. If not provided they default to the value returned by :meth:`alliance_platform.auth.resolve_perm_name`
     for the action 'create' and 'update' respectively. To disable checking permissions, you can pass ``None`` - this will
     mean any user, including anonymous users, can generate upload urls.
 
-    When using django forms :class:`~alliance_platform.storage.fields.async_file.AsyncFileFormField` provides a widget for handling
-    the upload from the frontend. This is the default ``formfield`` provided by :class:`~alliance_platform.storage.fields.async_file.AsyncFileField`.
+    When using django forms :class:`~alliance_platform.storage.async_uploads.forms.AsyncFileField` provides a widget for handling
+    the upload from the frontend. This is the default ``formfield`` provided by :class:`~alliance_platform.storage.async_uploads.models.AsyncFileField`.
 
     .. note:: The key for the file is stored in the database as a CharField and as such has a max_length. The default
         for this is ``500``. This must be sufficient to accommodate the temporary file value which looks something like
-        ``async-temp-files/2021/03/03/fVy5cSVBQpOb-test.png`` by default (see :meth:`~alliance_platform.storage.base.AsyncUploadStorage.generate_temporary_path`
+        ``async-temp-files/2021/03/03/fVy5cSVBQpOb-test.png`` by default (see :meth:`~alliance_platform.storage.async_uploads.storage.AsyncUploadStorage.generate_temporary_path`
         for how to customise this).
 
         Anything after the ``-`` will be truncated if necessary for the purposes of the temporary key. The actual filename
@@ -228,39 +226,39 @@ class AsyncFileMixin(AsyncFileMixinProtocol):
 
     The flow for how this works is as follows:
 
-        1) When a form is rendered on the frontend (eg. using :class:`~alliance_platform.storage.fields.async_file.AsyncFileFormField`) it
-           knows the ``async_field_id`` from the registry and ``generate_upload_url`` which is to the :class:`~alliance_platform.storage.views.GenerateUploadUrlView`
+        1) When a form is rendered on the frontend (eg. using :class:`~alliance_platform.storage.async_uploads.forms.AsyncFileField`) it
+           knows the ``async_field_id`` from the registry and ``generate_upload_url`` which is to the :class:`~alliance_platform.storage.async_uploads.views.GenerateUploadUrlView`
            view for the registry. This URL will be used to generated a specific URL for each file to upload to.
 
         2) When an upload occurs on the frontend it first hits ``generate_upload_url`` and passes the ``async_field_id``, the filename
-           and optionally an ``instance_id`` if it's an update for an existing record. :class:`~alliance_platform.storage.views.GenerateUploadUrlView`
+           and optionally an ``instance_id`` if it's an update for an existing record. :class:`~alliance_platform.storage.async_uploads.views.GenerateUploadUrlView`
            looks up the registry for the ``async_field_id`` to get the field and checks permissions on it (``perm_update`` if
            ``instance_id`` is passed otherwise ``perm_create``). If the permission check passes it will then create a
-           :class:`~alliance_platform.storage.models.AsyncTempFile` record to track the ``filename`` passed in and a :meth:`generated key <alliance_platform.storage.base.AsyncUploadStorage.generate_temporary_path>`
+           :class:`~alliance_platform.storage.async_uploads.models.AsyncTempFile` record to track the ``filename`` passed in and a :meth:`generated key <alliance_platform.storage.async_uploads.storage.AsyncUploadStorage.generate_temporary_path>`
            that will be used to upload the file to a temporary location on the storage backend. The view will return an
-           upload url by calling :meth:`~alliance_platform.storage.base.AsyncUploadStorage.generate_upload_url` along with the generated ``key``.
+           upload url by calling :meth:`~alliance_platform.storage.async_uploads.storage.AsyncUploadStorage.generate_upload_url` along with the generated ``key``.
 
         3) The frontend will receive the ``upload_url`` and ``key`` and proceed to upload to it. When the form is submitted
-           and saved the ``key`` (the value returned from :meth:`generated key <alliance_platform.storage.base.AsyncUploadStorage.generate_temporary_path>`)
+           and saved the ``key`` (the value returned from :meth:`generated key <alliance_platform.storage.async_uploads.storage.AsyncUploadStorage.generate_temporary_path>`)
            is the what will be stored in the database.
 
         4) On save a ``post_save`` signal will be called which will check if the file needs to be moved to a permanent
-           location. It does this by calling :meth:`~alliance_platform.storage.base.AsyncUploadStorage.is_temporary_path` and
-           :meth:`~alliance_platform.storage.base.AsyncUploadStorage.move_file` is called to move the file to its permanent
+           location. It does this by calling :meth:`~alliance_platform.storage.async_uploads.storage.AsyncUploadStorage.is_temporary_path` and
+           :meth:`~alliance_platform.storage.async_uploads.storage.AsyncUploadStorage.move_file` is called to move the file to its permanent
            location at which point the ``AsyncTempFile`` is deleted.
 
-        5) If an upload occurs but the form isn't submitted :class:`~alliance_platform.storage.models.AsyncTempFile` will be created
+        5) If an upload occurs but the form isn't submitted :class:`~alliance_platform.storage.async_uploads.models.AsyncTempFile` will be created
            but never deleted. The :class:`~alliance_platform.storage.management.commands.cleanup_async_temp_files` should be run
            periodically to clean these up.
 
     The url returned when accessing the ``url`` property, eg. ``model.file_field.url`` will always return the URL for
-    :class:`~alliance_platform.storage.views.DownloadRedirectView`. This view will check the user has permission to download the
+    :class:`~alliance_platform.storage.async_uploads.views.DownloadRedirectView`. This view will check the user has permission to download the
     file by checking the ``perm_download`` permission. If not provided this defaults to the value returned by
     :meth:`alliance_platform.auth.resolve_perm_name` with the action "detail" (ie. if they have permission to view the record
     they can download the file). This view will then redirect to the URL provided by
-    :meth:`~alliance_platform.storage.base.AsyncUploadStorage.generate_download_url`.
+    :meth:`~alliance_platform.storage.async_uploads.storage.AsyncUploadStorage.generate_download_url`.
 
-    .. note:: You must you this with a storage class that implements :class:`~alliance_platform.storage.base.AsyncUploadStorage`.
+    .. note:: You must you this with a storage class that implements :class:`~alliance_platform.storage.async_uploads.storage.AsyncUploadStorage`.
         Either set DEFAULT_FILE_STORAGE to a class (eg. :class:`~alliance_platform.storage.s3.S3AsyncUploadStorage`) or pass an
         instance in the ``storage`` kwarg.
 
@@ -269,17 +267,17 @@ class AsyncFileMixin(AsyncFileMixinProtocol):
     Codegen will create this as a ``AsyncFileField`` or ``AsyncImageField`` on the frontend Presto model.
     ``getWidgetForField`` will map these fields to the ``UploadWidget``.
 
-    See :class:`~alliance_platform.storage.fields.async_file.AsyncFileField` and :class:`~alliance_platform.storage.fields.async_file.AsyncFileField`
+    See :class:`~alliance_platform.storage.async_uploads.models.AsyncFileField` and :class:`~alliance_platform.storage.async_uploads.models.AsyncFileField`
     for Field classes that use this.
 
-    .. warning:: User input should be assigned using the :class:`~alliance_platform.storage.fields.async_file.AsyncFileInputData` class
+    .. warning:: User input should be assigned using the :class:`~alliance_platform.storage.async_uploads.models.AsyncFileInputData` class
         which has validation to disallow things like moving the file by changing the key that passed from the frontend. This
-        is handled for you when using :class:`~alliance_platform.storage.fields.async_file.AsyncFileField`,
-        :class:`~alliance_platform.storage.fields.async_file.AsyncImageField`, :class:`~alliance_platform.storage.drf.serializer.AsyncFileField`
+        is handled for you when using :class:`~alliance_platform.storage.async_uploads.models.AsyncFileField`,
+        :class:`~alliance_platform.storage.async_uploads.models.AsyncImageField`, :class:`~alliance_platform.storage.drf.serializer.AsyncFileField`
         or :class:`~alliance_platform.storage.drf.serializer.AsyncImageField`
 
     If a file cannot be moved for any reason (eg. any exception is raised) then the error will be saved in the
-    ``error`` field on the :class:`~alliance_platform.storage.models.AsyncTempFile` record and the temporary key will be retained as
+    ``error`` field on the :class:`~alliance_platform.storage.async_uploads.models.AsyncTempFile` record and the temporary key will be retained as
     the field value. The error will also be logged to the ``alliance_platform.storage`` logger. This cannot be resolved automatically
     and so requires manual intervention to cleanup.
 
@@ -330,7 +328,7 @@ class AsyncFileMixin(AsyncFileMixinProtocol):
     def contribute_to_class(self, cls, name: str, private_only: bool = False):
         super().contribute_to_class(cls, name, private_only=private_only)  # type: ignore[safe-super]
         if not cls._meta.abstract and self.model._meta.app_config:
-            field = cast(AsyncFileField, self)
+            field = cast(_AsyncFileField, self)
             AsyncFileModelRegistry.register_field(cls, field)
             self.async_field_registry.register_field(field)
             if self.perm_update is _UNSET:
@@ -388,7 +386,7 @@ class AsyncFileInputData:
     """The data we receive from the frontend gets converted to an instance of this
 
     :code:`width` and :code:`height` can optionally be included for images to avoid having
-    to calculate dimensions manually when an :class:`~alliance_platform.storage.fields.async_file.AsyncImageField` uses the
+    to calculate dimensions manually when an :class:`~alliance_platform.storage.async_uploads.models.AsyncImageField` uses the
     ``width_field`` and ``height_field`` options.
     """
 
@@ -396,20 +394,20 @@ class AsyncFileInputData:
     key: str
     # The original name of the file
     name: str
-    # The width of the image (optional - only applicable to :class:`~alliance_platform.storage.fields.async_file.AsyncImageField`)
+    # The width of the image (optional - only applicable to :class:`~alliance_platform.storage.async_uploads.models.AsyncImageField`)
     width: int | None = None
-    # The height of the image (optional - only applicable to :class:`~alliance_platform.storage.fields.async_file.AsyncImageField`)
+    # The height of the image (optional - only applicable to :class:`~alliance_platform.storage.async_uploads.models.AsyncImageField`)
     height: int | None = None
     # Only set if there was an upload error on the frontend
     error: str | None = None
 
     def update_dimension_cache(self, field: models.Field):
-        """Updates the dimension_cache on :class:`~alliance_platform.storage.fields.async_file.AsyncImageField`"""
+        """Updates the dimension_cache on :class:`~alliance_platform.storage.async_uploads.models.AsyncImageField`"""
 
         width = self.width
         height = self.height
         if width and height:
-            # See AsyncImageField.dimension_cache. We do this rather than
+            # See _AsyncImageField.dimension_cache. We do this rather than
             # assigning directly to the width + height fields on instance
             # as when creating a new instance by passing in kwargs then
             # the width + height fields may be set to the the defaults
@@ -433,7 +431,8 @@ class AsyncFileInputData:
     def create_from_user_input(cls, values: dict[str, str | int]):
         """Create instance from user input.
 
-        Throws if invalid keys present in ``values``. ``url`` is accepted but ignored."""
+        Throws if invalid keys present in ``values``. ``url`` is accepted but ignored.
+        """
         data: AsyncFileInputDataKwargs = {}
         valid_fields = ["key", "name", "width", "height", "error"]
         # We return "url" in format_value but it's not something
@@ -450,7 +449,7 @@ class AsyncFileInputData:
             raise ValidationError(f"Invalid input keys received: {', '.join(invalid_input)}")
         return cls(**data)
 
-    def validate_key(self, field: AsyncFileField, existing_value: str | AsyncFieldFile | None):
+    def validate_key(self, field: _AsyncFileField, existing_value: str | AsyncFieldFile | None):
         """Validate the key should be accepted
 
         This checks to make sure the key is a temporary path or matches the existing field
@@ -470,196 +469,6 @@ class AsyncFileInputData:
                     f"with a validation error."
                 )
                 raise ValidationError({field.name: "Invalid upload received"})
-
-
-# This does not extend FileInput as we don't actually receive a file
-# from the frontend. Instead we get a string which represents the key
-# for the storage backend.
-class AsyncFileInput(Input):
-    """Input for handling async uploads
-
-    This handles the submission value from UploadWidget and converts it to an instance
-    of :class:`~alliance_platform.storage.fields.async_file.AsyncFileInputData`. This is then handled
-    on the descriptor classes for :class:`~alliance_platform.storage.fields.async_file.AsyncFileField`
-    and :class:`~alliance_platform.storage.fields.async_file.AsyncImageField`.
-
-    To customise the widget rendered on the frontend you can override the template
-    ``alliance_platform/storage/widgets/async_file_input.html``.
-    """
-
-    input_type = "async-file"
-    template_name = "alliance_platform/storage/widgets/async_file_input.html"
-
-    # This is needed to generate URL to file in case where we only
-    # have the AsyncTempFile (see `format_value`)
-    storage: AsyncUploadStorage
-
-    def get_context(self, name, value, attrs):
-        context = super().get_context(name, value, attrs)
-
-        # Resolving the upload url needs to be delayed until render so that the
-        # GenerateUploadUrl view has a chance to attach to the registry. Resolve
-        # that here.
-        async_field_registry = context["widget"]["attrs"].pop("async_field_registry", None)
-        if not async_field_registry:
-            raise ValueError("AsyncFileInput expects a 'async_field_registry' to be passed in widget attrs")
-
-        context["widget"]["attrs"]["generate_upload_url"] = reverse(async_field_registry.attached_view)
-
-        return context
-
-    def format_value(self, value):
-        """Given a value return it in serialized format expected by frontend
-
-        Frontend expects a json string like:
-
-            {"key": "/some/storage/location/test.png", "name": "test.png" }
-        """
-        if not value:
-            return None
-        if isinstance(value, AsyncFileInputData):
-            # If a submission occurs but isn't successful (eg. validation fails on another
-            # field) we end up with a string which should match the key on AsyncTempFile
-            try:
-                temp_file = AsyncTempFile.objects.get(key=value.key)
-                return json.dumps(
-                    {
-                        "key": temp_file.key,
-                        "name": temp_file.original_filename,
-                        "url": self.storage.url(temp_file.key),
-                    }
-                )
-            except AsyncTempFile.DoesNotExist:
-                return json.dumps(
-                    {
-                        "key": value.key,
-                        "name": value.name,
-                    }
-                )
-        try:
-            url = value.url
-        except Exception:
-            # url isn't critical; if it fails ignore it and frontend may just not be able
-            # to show a thumbnail
-            url = None
-        return json.dumps({"key": value.name, "name": value.name.split("/")[-1], "url": url})
-
-    def value_from_datadict(self, data, files, name):
-        """This expects to receive a valid json string that can be used to instantiate AsyncFileInputData
-
-        If invalid JSON is received this is handled by the ``AsyncFileInputDataValidator`` which will
-        raise a ValidationError. We load the json here rather than in the field ``clean`` so that we have
-        the ``AsyncFileInputData`` in ``format`` for the case where there's a ValidationError and form
-        re-renders.
-
-        AsyncFileDescriptor & AsyncImageDescriptor handle extracting the ``key`` from this which is
-        what is set against the field on the model.
-
-        We need the AsyncFileInputData to pass across extra details that we may need - eg. width & height
-        for image fields
-        """
-        value = data.get(name)
-        try:
-            value = json.loads(value)
-            if value:
-                return AsyncFileInputData.create_from_user_input(value)
-            return value
-        except JSONDecodeError:
-            return value
-
-
-class AsyncFileInputDataLengthValidator(validators.MaxLengthValidator):
-    """Validate length of ``key`` on an AsyncFileInputData
-
-    FileField passes ``max_length`` to the form field and the field is, at the database level,
-    a char field. The normal FileField handles this internally but we can't extend forms.FileField
-    as it expects to be handling File submissions which we aren't doing here. This validator
-    just extracts the ``key`` and validates the length of that which is what gets written to the
-    database.
-    """
-
-    def clean(self, x):
-        if isinstance(x, AsyncFileInputData):
-            # We want to check the length of the key which is what gets stored in db
-            return len(x.name)
-        return len(x)
-
-
-@deconstructible
-class AsyncFileInputDataValidator:
-    """AsyncFileInputData has an error key that can be sent from the frontend - this validator just checks that"""
-
-    def __call__(self, value):
-        if isinstance(value, AsyncFileInputData):
-            if value.error:
-                raise ValidationError(f"There was an unexpected error: {value.error}")
-        else:
-            raise ValidationError("Bad input for file field")
-
-
-class AsyncFileFormField(forms.Field):
-    """Form field that renders a :class:`~alliance_platform.storage.fields.async_file.AsyncFileInput`
-
-    This is the default form field for :class:`alliance_platform.storage.fields.async_file.AsyncFileField`
-    """
-
-    widget = AsyncFileInput
-
-    async_field_registry: AsyncFieldRegistry
-    async_field_id: str
-
-    def __init__(
-        self,
-        *args,
-        async_field_registry: AsyncFieldRegistry,
-        async_field_id: str,
-        storage: AsyncUploadStorage,
-        max_length=None,
-        **kwargs,
-    ):
-        """
-
-        Args:
-            *args: Any additional arguments to pass through to :class:`django.forms.Field`
-            async_field_registry: The async field registry that should used on the frontend to create
-                unique upload urls for files. This typically comes from :code:`field.async_field_registry` where :code:`field`
-                is an :class:`~alliance_platform.storage.fields.async_file.AsyncFileField`.
-            async_field_id: The string ID of the field used in the registry. Generated with ``field.async_field_registry.generate_id(field)``.
-            storage: The storage class. This comes from the field ``field.storage``.
-            max_length: Max length for the filename
-            **kwargs:ny additional keyword arguments to pass through to :class:`django.forms.Field`
-        """
-        self.async_field_registry = async_field_registry
-        self.async_field_id = async_field_id
-        # This comes from FileField
-        self.max_length = max_length
-        super().__init__(*args, **kwargs)
-
-        # No way to pass args to a widget with how forms.Field work.. just
-        # set it after creation
-        self.widget.storage = storage
-
-        if max_length is not None:
-            self.validators.append(AsyncFileInputDataLengthValidator(int(max_length)))
-        self.validators.append(AsyncFileInputDataValidator())
-
-    def widget_attrs(self, widget):
-        attrs = super().widget_attrs(widget)
-        attrs["async_field_registry"] = self.async_field_registry
-        attrs["async_field_id"] = self.async_field_id
-        return attrs
-
-
-class AsyncImageFormField(AsyncFileFormField):
-    widget = AsyncFileInput
-
-    def widget_attrs(self, widget):
-        attrs = super().widget_attrs(widget)
-        if "accept" not in attrs:
-            attrs.setdefault("accept", "image/*")
-        if "list_type" not in attrs:
-            attrs.setdefault("list_type", "picture")
-        return attrs
 
 
 class FieldFileMixin:
@@ -686,158 +495,50 @@ class FieldFileMixin:
         return f"{url}?{query}"
 
 
-class AsyncFileDescriptor(FileDescriptor):
-    """Extracts the ``key`` from AsyncFileInputData to be set against instance"""
-
-    def __set__(self, instance, value):
-        if isinstance(value, AsyncFileInputData):
-            value.validate_key(cast(AsyncFileField, self.field), instance.__dict__.get(self.field.name))
-            value = value.key
-        super().__set__(instance, value)
-
-
 class AsyncFieldFile(FieldFileMixin, FieldFile):
     pass
 
 
-class AsyncFileField(AsyncFileMixin, FileField):  # type: ignore[misc] # mypy dont like the fact that we're using AsyncUploadStorage and report it as conflict w/ FileField
-    """A FileField that works with AsyncUploadStorage to directly upload somewhere (eg. S3) from the frontend
+class AsyncTempFile(models.Model):
+    """Model to track files that are being uploaded to a temporary location
 
-    See :class:`~alliance_platform.storage.fields.async_file.AsyncFileMixin` for more details on how this field works.
+    :class:`~alliance_platform.storage.async_uploads.views.GenerateUploadUrlView` is used to generate a URL to directly upload
+    a file to. When this URL is generated an :class:`AsyncTempFile` is created to track the new
+    key that is used (eg. /temp/2020/01/04/abc123-myfile.png), the original filename (eg. myfile.png)
+    and the specific field it came from (this is done via :class:`alliance_platform.storage.async_uploads.registry.AsyncFieldRegistry`).
+
+    Once a file has been uploaded and the form saved the key recorded here will be saved against the
+    underlying file field (either :class:`~alliance_platform.storage.async_uploads.models.AsyncFileField` or :class:`~alliance_platform.storage.async_uploads.models.AsyncImageField`)
+    on the target model which will check if that key is a temporary file using an :meth:`~alliance_platform.storage.async_uploads.storage.AsyncUploadStorage.is_temporary_path`. If so the
+    file will be moved to its permanent location using :meth:`~alliance_platform.storage.async_uploads.storage.AsyncUploadStorage.move_file`
+    and the :class:`AsyncTempFile` record will have the ``moved_to_location`` value set. See :class:`~alliance_platform.storage.async_uploads.models.AsyncFileMixin`
+    for where this happens and more details.
+
+    You must run the :class:`~alliance_platform.storage.management.commands.cleanup_async_temp_files` command periodically to
+    cleanup this table. This handles two cases: the success case where ``moved_to_location`` is set but the record is
+    being kept around for a while to detect duplicate submissions, and the other case where upload occurred on the frontend
+    but the form was never submitted and the file was never moved.
     """
 
-    attr_class = AsyncFieldFile
-    descriptor_class = AsyncFileDescriptor
+    created_at = models.DateTimeField(auto_now_add=True)
+    original_filename = models.TextField()
+    # This is the temporary location the file will be saved to and is generated using AsyncUploadStorage.generate_temporary_path(filename)
+    key = models.CharField(max_length=500)
+    content_type = models.ForeignKey(ContentType, on_delete=models.CASCADE)
+    field_name = models.CharField(max_length=100)
+    # This will be set once the file has been moved to its permanent location. We store this so we can detect duplicate submissions and handle them appropriately.
+    moved_to_location = models.TextField(blank=True)
+    # This will be set if there was an error trying to move the file. When this is set cleanup_async_temp_files will not remove the record.
+    error = models.TextField(null=True)
 
-    def formfield(self, **kwargs):
-        return super().formfield(
-            **{
-                "form_class": AsyncFileFormField,
-                "async_field_id": self.async_field_registry.generate_id(self),
-                "async_field_registry": self.async_field_registry,
-                "storage": self.storage,
-                **kwargs,
-            }
+    @classmethod
+    def create_for_field(cls, field: _AsyncFileField | _AsyncImageField, filename: str) -> AsyncTempFile:
+        return AsyncTempFile.objects.create(
+            original_filename=filename,
+            content_type=ContentType.objects.get_for_model(field.model),
+            field_name=field.name,
+            key=field.storage.generate_temporary_path(filename, max_length=field.max_length),
         )
 
-
-class AsyncImageDescriptor(FileDescriptor):
-    """Extracts the ``key`` from AsyncFileInputData to be set against instance & handles width/height
-
-    If width & height are present they are set on the field and handled by
-    :meth:`alliance_platform.storage.fields.async_file.AsyncImageField.update_dimension_fields`.
-    """
-
-    def __set__(self, instance, value):
-        if isinstance(value, AsyncFileInputData):
-            value.validate_key(cast(AsyncFileField, self.field), instance.__dict__.get(self.field.name))
-            value.update_dimension_cache(self.field)
-            value = value.key
-            super().__set__(instance, value)
-        else:
-            # From ImageFileDescriptor:
-            # To prevent recalculating image dimensions when we are instantiating
-            # an object from the database (bug #11084), only update dimensions if
-            # the field had a value before this assignment.  Since the default
-            # value for FileField subclasses is an instance of field.attr_class,
-            # previous_file will only be None when we are called from
-            # Model.__init__().  The ImageField.update_dimension_fields method
-            # hooked up to the post_init signal handles the Model.__init__() cases.
-            # Assignment happening outside of Model.__init__() will trigger the
-            # update right here.
-
-            # We only do this if we don't have an AsyncFileInputData as
-            # that contains the width / height submitted from the
-            # frontend and so avoids potentially downloading the file
-            # to get the width/height.
-            previous_file = instance.__dict__.get(self.field.name)
-            super().__set__(instance, value)
-            # Don't compare value != previous_file as it doesn't guarantee the underlying
-            # file hasn't changed (eg. the key could be set to same filename again after
-            # a new upload but be a different file)
-            if previous_file is not None:
-                cast(AsyncImageField, self.field).update_dimension_fields(instance, force=True)
-
-
-class AsyncImageFieldFile(FieldFileMixin, ImageFieldFile):
-    pass
-
-
-class AsyncImageField(AsyncFileMixin, ImageField):  # type: ignore[misc] # mypy dont like the fact that we're using AsyncUploadStorage and report it as conflict w/ FileField
-    """A ImageField that works with AsyncUploadStorage to directly upload somewhere (eg. S3) from the frontend
-
-    This supports ``width_field`` and ``height_field`` in two ways
-
-    1) (preferred) The frontend passes the width & height with the form submission. This is supported with the
-    :class:`~alliance_platform.storage.fields.async_file.AsyncImageFormField` form field (the default). This works in
-    conjunction with the ``UploadWidget`` on the frontend.
-
-    2) (slow) Same behaviour as :class:`~django.db.models.ImageField` which requires the file to be downloaded
-    and processed with Pillow.
-
-    See :class:`~alliance_platform.storage.fields.async_file.AsyncFileMixin` for more details on how this field works.
-    """
-
-    descriptor_class = AsyncImageDescriptor
-    attr_class = AsyncImageFieldFile
-
-    # Cache dimensions as (width, height) from AsyncImageDescriptor. Used in update_dimension_fields
-    dimension_cache: tuple[int, int] | None
-
-    # Don't check if Pillow is installed. This is used to get dimensions of images.
-    _supress_pillow_check: bool
-
-    width_field: str | None
-    height_field: str | None
-
-    def __init__(self, *args, suppress_pillow_check=False, **kwargs):
-        super().__init__(*args, **kwargs)
-        self._suppress_pillow_check = suppress_pillow_check
-        self.dimension_cache = None
-
-    def check(self, **kwargs):
-        if self._suppress_pillow_check:
-            return super(ImageField, self).check(**kwargs)
-        return super().check(**kwargs)
-
-    def update_dimension_fields(self, instance, force=False, *args, **kwargs):
-        # If we have a cache use that otherwise fall back to default. Note that
-        # the default will open the file to work out dimensions and so with
-        # remote backends like S3 would result in downloading the file first.
-        if self.dimension_cache:
-            has_dimension_fields = self.width_field or self.height_field
-            if not has_dimension_fields or self.attname not in instance.__dict__:
-                return
-            width, height = self.dimension_cache
-            self.dimension_cache = None
-            if self.width_field:
-                setattr(instance, self.width_field, width)
-            if self.height_field:
-                setattr(instance, self.height_field, height)
-        else:
-            super().update_dimension_fields(instance, force, *args, **kwargs)
-
-    def before_move_file(self, instance):
-        """Prior to moving file populate the dimensions cache
-
-        This is so the moved file retains same dimensions as prior to move without
-        needing to recalculate it (ie. it's the same file, just in a different location).
-        """
-        has_dimension_fields = self.width_field and self.height_field
-        if has_dimension_fields and not self.dimension_cache:
-            width = getattr(
-                instance, cast(str, self.width_field)
-            )  # here -> has_dimension_fields -> str | None is str
-            height = getattr(instance, cast(str, self.height_field))
-            self.dimension_cache = (width, height)
-
-    def formfield(self, **kwargs):
-        return super().formfield(
-            **{
-                "form_class": AsyncImageFormField,
-                "async_field_id": self.async_field_registry.generate_id(self),
-                "async_field_registry": self.async_field_registry,
-                "storage": self.storage,
-                **kwargs,
-            }
-        )
+    class Meta:
+        db_table = "alliance_platform_storage_async_temp_file"
