+++ packages/ap-storage/alliance_platform/storage/async_uploads/storage/s3.py
@@ -1,7 +1,7 @@
 from typing import Any
 
-from alliance_platform.storage.base import AsyncUploadStorage
-from alliance_platform.storage.base import GenerateUploadUrlResponse
+from alliance_platform.storage.async_uploads.storage.base import AsyncUploadStorage
+from alliance_platform.storage.async_uploads.storage.base import GenerateUploadUrlResponse
 from alliance_platform.storage.settings import ap_storage_settings
 from django.core.exceptions import ImproperlyConfigured
