+++ packages/ap-storage/alliance_platform/storage/async_uploads/views/filesystem.py
@@ -1,10 +1,10 @@
 from typing import Any
 from typing import cast
 
-from alliance_platform.storage.filesystem import FileSystemAsyncUploadStorage
-from alliance_platform.storage.registry import default_async_field_registry
+from alliance_platform.storage.async_uploads.filesystem.upload_storage import FileSystemAsyncUploadStorage
+from alliance_platform.storage.async_uploads.registry import default_async_field_registry
+from alliance_platform.storage.async_uploads.views import ViewProtocol
 from alliance_platform.storage.settings import ap_storage_settings
-from alliance_platform.storage.views import ViewProtocol
 from django.core.files.uploadedfile import UploadedFile
 from django.core.signing import BadSignature
 from django.core.signing import SignatureExpired
@@ -20,19 +20,19 @@ from django.views.decorators.csrf import csrf_exempt
 
 @method_decorator(csrf_exempt, name="dispatch")
 class FileSystemAsyncStorageUploadView(View):
-    """View to handle a direct upload to the filesystem when using :class:`~alliance_platform.storage.filesystem.FileSystemAsyncUploadStorage`.
+    """View to handle a direct upload to the filesystem when using :class:`~alliance_platform.storage.async_uploads.storage.FileSystemAsyncUploadStorage`.
 
-    Each view is tied to a specific registry which you can specify in the ``registry`` kwarg (defaults to :data:`~alliance_platform.storage.registry.default_async_field_registry`).
+    Each view is tied to a specific registry which you can specify in the ``registry`` kwarg (defaults to :data:`~alliance_platform.storage.async_uploads.registry.default_async_field_registry`).
 
-    Note that this view does works off of signed data generated by :class:`~alliance_platform.storage.filesystem.FileSystemAsyncUploadStorage.generate_upload_url`,
+    Note that this view does works off of signed data generated by :class:`~alliance_platform.storage.async_uploads.storage.FileSystemAsyncUploadStorage.generate_upload_url`,
     and does not itself check any permissions. It does verify the signed data, including enforcing an expiry based on
     :data:`~alliance_platform.storage.settings.AlliancePlatformStorageSettings.UPLOAD_URL_EXPIRY`. This is similar to a
     signed URL generated by the S3 or Azure backends.
 
-    The permission checking is done in :class:`~alliance_platform.storage.views.GenerateUploadUrlView`, which
-    calls :class:`~alliance_platform.storage.filesystem.FileSystemAsyncUploadStorage.generate_upload_url`.
+    The permission checking is done in :class:`~alliance_platform.storage.async_uploads.views.GenerateUploadUrlView`, which
+    calls :class:`~alliance_platform.storage.async_uploads.storage.FileSystemAsyncUploadStorage.generate_upload_url`.
 
-    The URL for this view is automatically registered when :class:`~alliance_platform.storage.filesystem.FileSystemAsyncUploadStorage`
+    The URL for this view is automatically registered when :class:`~alliance_platform.storage.async_uploads.storage.FileSystemAsyncUploadStorage`
     is used if you have followed the :ref:`register-urls` guide.
     """
 
@@ -82,19 +82,19 @@ class FileSystemAsyncStorageUploadView(View):
 
 
 class FileSystemAsyncStorageDownloadView(View):
-    """View to handle a direct download from the filesystem when using :class:`~alliance_platform.storage.filesystem.FileSystemAsyncUploadStorage`.
+    """View to handle a direct download from the filesystem when using :class:`~alliance_platform.storage.async_uploads.storage.FileSystemAsyncUploadStorage`.
 
-    Each view is tied to a specific registry which you can specify in the ``registry`` kwarg (defaults to :data:`~alliance_platform.storage.registry.default_async_field_registry`).
+    Each view is tied to a specific registry which you can specify in the ``registry`` kwarg (defaults to :data:`~alliance_platform.storage.async_uploads.registry.default_async_field_registry`).
 
-    Note that this view does works off of signed data generated by :class:`~alliance_platform.storage.filesystem.FileSystemAsyncUploadStorage.generate_download_url`,
+    Note that this view does works off of signed data generated by :class:`~alliance_platform.storage.async_uploads.storage.FileSystemAsyncUploadStorage.generate_download_url`,
     and does not itself check any permissions. It does verify the signed data, including enforcing an expiry based on
     :data:`~alliance_platform.storage.settings.AlliancePlatformStorageSettings.DOWNLOAD_URL_EXPIRY`. This is similar to a
     signed URL generated by the S3 or Azure backends.
 
-    The permission checking is done in :class:`~alliance_platform.storage.views.GenerateUploadUrlView`, which
-    calls :class:`~alliance_platform.storage.filesystem.FileSystemAsyncUploadStorage.generate_upload_url`.
+    The permission checking is done in :class:`~alliance_platform.storage.async_uploads.views.GenerateUploadUrlView`, which
+    calls :class:`~alliance_platform.storage.async_uploads.storage.FileSystemAsyncUploadStorage.generate_upload_url`.
 
-    The URL for this view is automatically registered when :class:`~alliance_platform.storage.filesystem.FileSystemAsyncUploadStorage`
+    The URL for this view is automatically registered when :class:`~alliance_platform.storage.async_uploads.storage.FileSystemAsyncUploadStorage`
     is used if you have followed the :ref:`register-urls` guide.
     """
