--- packages/ap-storage/alliance_platform/storage/fields/__init__.py
@@ -1,1 +1,0 @@
-__all__ = ["AsyncFileField"]
