+++ packages/ap-storage/docs/usage.rst
@@ -1,7 +1,7 @@
 Usage
 =====
 
-To use these fields you must be using an :class:`~alliance_platform.storage.base.AsyncUploadStorage` storage class at either
+To use these fields you must be using an :class:`~alliance_platform.storage.async_uploads.storage.AsyncUploadStorage` storage class at either
 the global level, or field level. To setup globally see the :doc:`installation` instructions.
 
 To use at a field level create a storage class and pass it to the field:
@@ -14,8 +14,8 @@ To use at a field level create a storage class and pass it to the field:
         file = AsyncFileField(storage=storage)
         image = AsyncImageField(storage=storage)
 
-In addition to that two views need to be registered. :class:`~alliance_platform.storage.views.GenerateUploadUrlView` is used to
-generate a URL for the frontend to upload to (eg. a signed S3 URL). :class:`~alliance_platform.storage.views.DownloadRedirectView`
+In addition to that two views need to be registered. :class:`~alliance_platform.storage.async_uploads.views.GenerateUploadUrlView` is used to
+generate a URL for the frontend to upload to (eg. a signed S3 URL). :class:`~alliance_platform.storage.async_uploads.views.DownloadRedirectView`
 is used to check a user has permissions on a file before redirecting to the download location (eg. a signed S3 URL). This
 is covered in the :doc:`installation` guide.
 
@@ -23,17 +23,17 @@ Usage
 #####
 
 All usages require a model - there's currently no support for uploading files without the result being saved
-to an :class:`~alliance_platform.storage.fields.async_file.AsyncFileField` or :class:`~alliance_platform.storage.fields.async_file.AsyncImageField`.
+to an :class:`~alliance_platform.storage.async_uploads.models.AsyncFileField` or :class:`~alliance_platform.storage.async_uploads.models.AsyncImageField`.
 
 The default template setup will work automatically for files in both django forms and Presto forms. Uploading
 is performed using the UploadWidget on the frontend (see UploadWidget.tsx).
 
-For a deeper understanding of how django forms work see :class:`~alliance_platform.storage.fields.async_file.AsyncFileFormField`
-and :class:`~alliance_platform.storage.fields.async_file.AsyncFileInput`.
+For a deeper understanding of how django forms work see :class:`~alliance_platform.storage.async_uploads.forms.AsyncFileField`
+and :class:`~alliance_platform.storage.async_uploads.forms.AsyncFileInput`.
 
 DRF integration requires you to use the DRF fields :class:`alliance_platform.storage.drf.serializer.AsyncFileField` or
 :class:`alliance_platform.storage.drf.serializer.AsyncFileField`. See :ref:`storage-serializer-fields` for how to set
 this as a default.
 
-Intermediate files are stored in the :class:`alliance_platform.storage.models.AsyncTempFile` table. This should be periodically
+Intermediate files are stored in the :class:`alliance_platform.storage.async_uploads.models.AsyncTempFile` table. This should be periodically
 cleaned up by running the :class:`~alliance_platform.storage.management.commands.cleanup_async_temp_files` command.
