+++ packages/ap-storage/test_alliance_platform_storage/models.py
@@ -1,6 +1,6 @@
-from alliance_platform.storage.fields.async_file import AsyncFileField
-from alliance_platform.storage.fields.async_file import AsyncImageField
-from alliance_platform.storage.registry import AsyncFieldRegistry
+from alliance_platform.storage.async_uploads.models import AsyncFileField
+from alliance_platform.storage.async_uploads.models import AsyncImageField
+from alliance_platform.storage.async_uploads.registry import AsyncFieldRegistry
 from django.contrib.auth.models import AbstractUser
 from django.db import models
 from django.db.models import CASCADE
