+++ packages/ap-storage/test_alliance_platform_storage/urls.py
@@ -1,4 +1,4 @@
-from alliance_platform.storage.registry import default_async_field_registry
+from alliance_platform.storage.async_uploads.registry import default_async_field_registry
 
 app_name = "test_alliance_platform_storage"
